+++ docs/change_log.rst
@@ -1,5 +1,11 @@
 更新记录
 ===========================
+2018.8.13 v1.1.0
+---------------------------
+* `#74 <https://github.com/chinapnr/fishbase/issues/74>`_, common, add function :meth:`fish_common.get_group_list_data`, doc and unittest;
+* `#89 <https://github.com/chinapnr/fishbase/issues/89>`_, common, add function :meth:`fish_common.get_sub_dict`, doc and unittest;
+
+
 2018.8.2 v1.0.16
 ---------------------------
 * `#87 <https://github.com/chinapnr/fishbase/issues/87>`_, date, add function :meth:`fish_date.GetRandomTime`, doc and unittest;